+++ org/address.go
@@ -14,8 +14,7 @@ import (
 // a postal or fiscal address.
 // Attribute names loosely based on the xCard file format.
 type Address struct {
-	// Internal ID used to identify the party inside a document.
-	UUID uuid.UUID `json:"uuid,omitempty" jsonschema:"title=UUID"`
+	uuid.Identify
 	// Useful identifier, such as home, work, etc.
 	Label string `json:"label,omitempty" jsonschema:"title=Label"`
 	// Box number or code for the post office box located at the address.