+++ schema/object_test.go
@@ -1,15 +1,21 @@
 package schema_test
 
 import (
-	"encoding/json"
 	"testing"
 
+	"github.com/invopop/gobl/bill"
+	"github.com/invopop/gobl/cal"
+	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/note"
+	"github.com/invopop/gobl/num"
+	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/schema"
 	"github.com/invopop/gobl/tax"
 	"github.com/invopop/gobl/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	_ "github.com/invopop/gobl"
 )
 
 // See also document tests performed in `gobl` package.
@@ -17,41 +23,76 @@ import (
 func TestObjectUUID(t *testing.T) {
 	tr := &tax.Regime{} // doesn't have a UUID field!
 	obj, err := schema.NewObject(tr)
-	require.NoError(t, err)
-	assert.Equal(t, uuid.Empty, obj.UUID())
+	assert.NoError(t, err)
+	assert.Empty(t, obj.UUID())
 
 	msg := &note.Message{
-		UUID:    uuid.V1(),
 		Title:   "just a test",
 		Content: "this is a test message",
 	}
-
 	obj, err = schema.NewObject(msg)
 	require.NoError(t, err)
+	assert.Empty(t, msg.UUID)
+	assert.Empty(t, obj.UUID())
 
+	msg = &note.Message{
+		Title:   "just a test",
+		Content: "this is a test message",
+	}
+	msg.UUID = uuid.V1()
+	obj, err = schema.NewObject(msg)
+	require.NoError(t, err)
 	assert.Equal(t, msg.UUID, obj.UUID())
 }
 
-func TestObjectInjectUUID(t *testing.T) {
-	tr := &tax.Regime{} // doesn't have a UUID field!
-	id := uuid.V1()
-	obj, err := schema.NewObject(tr)
+func TestObjectCalculate(t *testing.T) {
+	inv := exampleInvoice()
+	obj, err := schema.NewObject(inv)
 	require.NoError(t, err)
-	assert.NotPanics(t, func() {
-		obj.InjectUUID(id)
-	})
 
-	msg := &note.Message{
-		Title:   "just a test",
-		Content: "this is a test message",
+	assert.Nil(t, inv.Totals)
+	assert.Empty(t, inv.UUID)
+	require.NoError(t, obj.Calculate())
+	assert.NotNil(t, inv.Totals)
+	assert.NotEmpty(t, inv.UUID)
+	assert.NotEmpty(t, obj.UUID())
+	assert.Equal(t, obj.UUID(), inv.UUID)
+}
+
+// exampleInvoice defines a simple invoice example pre-calculations.
+func exampleInvoice() *bill.Invoice {
+	return &bill.Invoice{
+		Code: "123TEST",
+		Tax: &bill.Tax{
+			PricesInclude: tax.CategoryVAT,
+		},
+		Supplier: &org.Party{
+			TaxID: &tax.Identity{
+				Country: l10n.ES,
+				Code:    "B98602642",
+			},
+		},
+		Customer: &org.Party{
+			TaxID: &tax.Identity{
+				Country: l10n.ES,
+				Code:    "54387763P",
+			},
+		},
+		IssueDate: cal.MakeDate(2022, 6, 13),
+		Lines: []*bill.Line{
+			{
+				Quantity: num.MakeAmount(1, 0),
+				Item: &org.Item{
+					Name:  "Item",
+					Price: num.MakeAmount(4320, 2),
+				},
+				Taxes: tax.Set{
+					{
+						Category: "VAT",
+						Percent:  num.NewPercentage(6, 2),
+					},
+				},
+			},
+		},
 	}
-	obj, err = schema.NewObject(msg)
-	require.NoError(t, err)
-
-	obj.InjectUUID(id)
-
-	assert.Equal(t, id, obj.UUID())
-	data, err := json.Marshal(obj)
-	require.NoError(t, err)
-	assert.Equal(t, `{"$schema":"https://gobl.org/draft-0/note/message","uuid":"`+id.String()+`","title":"just a test","content":"this is a test message"}`, string(data))
 }